+++ src/nary_order_unary_transformation_method2.c
@@ -1239,12 +1239,12 @@ static STRING *_CreateNewReactionID( REACTION *newReaction ) {
         if( IsLogicalSpecies( species ) ) {
             sprintf( buf, "Deg_%s__%X", 
                      GetCharArrayOfString( GetOriginalSpeciesName( (LOGICAL_SPECIES*)species ) ), 
-                     (int)newReaction );
+                     newReaction );
         }
         else {
             sprintf( buf, "Deg_%s__%X", 
                      GetCharArrayOfString( GetSpeciesNodeName( species ) ), 
-                     (int)newReaction );
+                     newReaction );
         }
         newID = CreateString( buf );
         return newID;
@@ -1257,12 +1257,12 @@ static STRING *_CreateNewReactionID( REACTION *newReaction ) {
         if( IsLogicalSpecies( species ) ) {
             sprintf( buf, "Prod_%s__%X", 
                      GetCharArrayOfString( GetOriginalSpeciesName( (LOGICAL_SPECIES*)species ) ), 
-                     (int)newReaction );
+                     newReaction );
         }
         else {
             sprintf( buf, "Prod_%s__%X", 
                      GetCharArrayOfString( GetSpeciesNodeName( species ) ), 
-                     (int)newReaction );
+                     newReaction );
         }
         newID = CreateString( buf );
         return newID;
