+++ src/nary_order_unary_transformation_method3.c
@@ -1266,12 +1266,12 @@ static STRING *_CreateNewReactionID( REACTION *newReaction ) {
         if( IsLogicalSpecies( species ) ) {
             sprintf( buf, "Deg_%s__%X", 
                      GetCharArrayOfString( GetOriginalSpeciesName( (LOGICAL_SPECIES*)species ) ), 
-                     (int)newReaction );
+                     newReaction );
         }
         else {
             sprintf( buf, "Deg_%s__%X", 
                      GetCharArrayOfString( GetSpeciesNodeName( species ) ), 
-                     (int)newReaction );
+                     newReaction );
         }
         newID = CreateString( buf );
         return newID;
@@ -1284,12 +1284,12 @@ static STRING *_CreateNewReactionID( REACTION *newReaction ) {
         if( IsLogicalSpecies( species ) ) {
             sprintf( buf, "Prod_%s__%X", 
                      GetCharArrayOfString( GetOriginalSpeciesName( (LOGICAL_SPECIES*)species ) ), 
-                     (int)newReaction );
+                     newReaction );
         }
         else {
             sprintf( buf, "Prod_%s__%X", 
                      GetCharArrayOfString( GetSpeciesNodeName( species ) ), 
-                     (int)newReaction );
+                     newReaction );
         }
         newID = CreateString( buf );
         return newID;
